06_emergence_mirage/requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0
scikit-learn>=1.2.0
openai>=1.0.0
anthropic>=0.18.0
tiktoken>=0.5.0
flask>=2.3.0
flask-socketio>=5.3.0
pyyaml>=6.0
tqdm>=4.65.0
pytest>=7.3.0
statsmodels>=0.14.0
